13066/tests/lib/requirements.txt
alabaster==0.7.8
ansible==2.9.12
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
autopep8==1.5.4
Babel==2.4.0
beautifulsoup4==4.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
constantly==15.1.0
convoy==0.4.4
crochet==1.4.0
cryptography==2.1.4
debtcollector==2.2.0
distro-info==0.18
Django==1.11.11
django-piston3==0.3rc2
djorm-pgarray==1.2
dnspython==1.15.0
docutils==0.14
FormEncode==1.3.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
importlib-metadata==1.7.0
incremental==16.10.1
isc==2.0
iso8601==0.1.11
Jinja2==2.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
keystoneauth1==4.2.1
language-selector==0.1
lxml==4.2.1
macaroonbakery==1.2.1
MarkupSafe==1.0
msgpack==1.0.0
netaddr==0.7.19
netifaces==0.10.4
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
os-service-types==1.7.0
oslo.config==8.3.1
oslo.i18n==5.0.0
oslo.serialization==4.0.0
oslo.utils==4.4.0
packaging==20.4
PAM==0.4.2
paramiko==2.0.0
pbr==3.1.1
petname==2.2
pexpect==4.2.1
Pillow==5.1.0
ply==3.11
prettytable==0.7.2
protobuf==3.0.0
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.6.0
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.3+ubuntu1
python-keystoneclient==4.1.0
python-mimeparse==0.1.4
python-seamicroclient==0.4.0
python-tx-tftp==0.2
pytz==2018.3
pyvmomi==6.5.0.2017.5.post1
pyxdg==0.25
PyYAML==5.3.1
requests==2.18.4
rfc3986==1.4.0
roman==2.0.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
SecretStorage==2.3.1
service-identity==16.0.0
simplejson==3.13.2
six==1.11.0
Sphinx==1.6.7
sqlparse==0.2.4
ssh-import-id==5.7
stevedore==3.2.0
Tempita==0.5.2
toml==0.10.1
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
uvloop==0.8.1
webencodings==0.5
wrapt==1.12.1
zipp==3.1.0
zope.interface==4.3.2
